Modificar subasta

+++ src/main/java/edu/eci/arsw/saleit/model/Articulo.java
@@ -128,7 +128,7 @@ public class Articulo {
         this.imagen = imagen;
     }
 
-    public long getCategoria() {
+    public int getCategoria() {
         return categoria;
     }
